+++ example/NonTerminate.hs
@@ -7,7 +7,7 @@ import Frontend
 nonterminate :: Ref Int -> Exp Int -> SSM ()
 nonterminate = box "nonterminate" ["x", "y"] $ \x y -> do
     v <- deref x
-    after (int 1) x (v +. (int 1))
+    after (uint64 1) x (v +. (int 1))
     wait [x]
     fork [nonterminate x (v + y)]
